+++ db/migrate/20210311142246_add_completed_at_to_analyses_and_evaluations.rb
@@ -1,0 +1,6 @@
+class AddCompletedAtToAnalysesAndEvaluations < ActiveRecord::Migration[6.1]
+  def change
+    add_column :analyses, :completed_at, :datetime, after: :performed_at
+    add_column :evaluations, :completed_at, :datetime, after: :performed_at
+  end
+end