+++ spec/support/shareds/views/analysis_view_shared.rb
@@ -11,6 +11,14 @@ shared_context '分析ジョブを作成する' do |total: nil, update_attribute
   end
 end
 
+shared_context '分析フォームオブジェクトを作成する' do |attribute = {}|
+  before(:all) do
+    @new_analysis = Analysis.new
+    @new_analysis.build_parameter
+    @index_form = Analyses::IndexForm.new(attribute)
+  end
+end
+
 shared_examples '分析画面共通テスト' do |expected: {}|
   it_behaves_like 'ヘッダーが表示されていること'
   it_behaves_like '分析情報入力フォームが表示されていること'
@@ -23,10 +31,35 @@ shared_examples '分析画面共通テスト' do |expected: {}|
 end
 
 shared_examples '分析情報入力フォームが表示されていること' do
-  it_behaves_like 'タイトルが表示されていること', 'レースを分析'
+  [
+    %w[登録 new-analysis],
+    %w[検索 search-form],
+  ].each do |tab_name, tab_href|
+    it "#{tab_name}タブが表示されていること" do
+      tab = @html.xpath("#{form_tab_xpath}/li/a[@href='##{tab_href}']")
+      is_asserted_by { tab.present? }
+      is_asserted_by { tab.text.strip == tab_name }
+    end
+  end
+
+  it '分析ジョブ登録フォームがアクティブになっていること' do
+    tab = @html.xpath("#{form_tab_xpath}/li[@class='active']/a[@href='#new-analysis']")
+    is_asserted_by { tab.present? }
+  end
+
+  it_behaves_like '分析ジョブ登録フォームが表示されていること'
+  it_behaves_like '分析ジョブ検索フォームが表示されていること'
+end
+
+shared_examples '分析ジョブ登録フォームが表示されていること' do
+  it 'タイトルが表示されていること' do
+    title = @html.xpath("#{register_form_panel_xpath}/h3")
+    is_asserted_by { title.present? }
+    is_asserted_by { title.text.strip == 'レースを分析' }
+  end
 
   it '必須項目の説明が表示されていること' do
-    description = @html.xpath("#{form_panel_xpath}/h4")
+    description = @html.xpath("#{register_form_panel_xpath}/h4")
     is_asserted_by { description.present? }
     is_asserted_by { description.text.strip == '* は必須項目です' }
   end
@@ -36,25 +69,25 @@ shared_examples '分析情報入力フォームが表示されていること' d
     %w[num_entry エントリー数],
   ].each do |param_name, label_name|
     it "#{label_name}入力フォームのラベルがあること" do
-      label = @html.xpath("#{input_xpath}/label[@for='analysis_#{param_name}']")
+      label = @html.xpath("#{register_input_xpath}/label[@for='analysis_#{param_name}']")
       is_asserted_by { label.present? }
-      is_asserted_by { label.text == label_name }
+      is_asserted_by { label.text.strip == label_name }
     end
 
     it "#{label_name}入力フォームがあること" do
-      input = @html.xpath("#{input_xpath}/input[@id='analysis_#{param_name}']")
+      input = @html.xpath("#{register_input_xpath}/input[@id='analysis_#{param_name}']")
       is_asserted_by { input.present? }
       is_asserted_by { input.attribute('value').nil? }
     end
   end
 
   it 'パラメーター入力フォームが閉じた状態で表示されていること' do
-    parameter_form_link = @html.xpath(input_xpath).search('a[@href="#parameter"]')
-    is_asserted_by { parameter_form_link.present? }
-    is_asserted_by { parameter_form_link.text.strip == 'パラメーター設定' }
+    link = @html.xpath(register_input_xpath).search('a[@href="#parameter"]')
+    is_asserted_by { link.present? }
+    is_asserted_by { link.text.strip == 'パラメーター設定' }
 
-    parameter_form_block = @html.xpath(form_xpath).search('div[@class="collapse"]')
-    is_asserted_by { parameter_form_block.present? }
+    block = @html.xpath(register_form_xpath).search('div[@class="collapse"]')
+    is_asserted_by { block.present? }
   end
 
   [
@@ -70,7 +103,7 @@ shared_examples '分析情報入力フォームが表示されていること' d
       form_id = "analysis_parameter_attributes_#{param_name}"
       label = parameter_form_block.search("label[@for='#{form_id}']")
       is_asserted_by { label.present? }
-      is_asserted_by { label.text == param_name }
+      is_asserted_by { label.text.strip == param_name }
 
       input = parameter_form_block.search("input[@id='#{form_id}']")
       is_asserted_by { input.present? }
@@ -103,10 +136,82 @@ shared_examples '分析情報入力フォームが表示されていること' d
     %w[reset リセット],
   ].each do |type, value|
     it "#{value}ボタンがあること" do
-      button = @html.xpath("#{form_xpath}/input[@type='#{type}']")
+      button = @html.xpath("#{register_form_xpath}/input[@type='#{type}']")
       is_asserted_by { button.present? }
       is_asserted_by { button.attribute('value').value == value }
     end
+  end
+end
+
+shared_examples '分析ジョブ検索フォームが表示されていること' do
+  it 'タイトルが表示されていること' do
+    title = @html.xpath("#{index_form_panel_xpath}/h3")
+    is_asserted_by { title.present? }
+    is_asserted_by { title.text.strip == 'ジョブを検索' }
+  end
+
+  it '学習データ数入力フォームのラベルがあること' do
+    label = @html.xpath("#{index_input_xpath}/label[@for='input-index-num_data']")
+    is_asserted_by { label.present? }
+    is_asserted_by { label.text.strip == '学習データ数' }
+  end
+
+  it '学習データ数入力フォームがあること' do
+    input = @html.xpath("#{index_input_xpath}/input[@id='input-index-num_data']")
+    is_asserted_by { input.present? }
+    is_asserted_by { input.attribute('value')&.value&.to_i == @index_form.num_data }
+  end
+
+  it 'パラメーター入力ラベルが表示されていること' do
+    label = @html.xpath("#{index_form_xpath}/div[@id='label-index-parameter']")
+    is_asserted_by { label.present? }
+    is_asserted_by { label.text.strip == 'パラメーター' }
+  end
+
+  %w[
+    max_depth
+    max_leaf_nodes
+    min_samples_leaf
+    min_samples_split
+    num_tree
+  ].each do |param_name|
+    it "#{param_name}入力フォームがあること" do
+      form_id = "input-index-#{param_name}"
+      label = @html.xpath("#{index_input_xpath}/label[@for='#{form_id}']")
+      is_asserted_by { label.present? }
+      is_asserted_by { label.text.strip == param_name }
+
+      input = @html.xpath("#{index_input_xpath}/input[@id='#{form_id}']")
+      is_asserted_by { input.present? }
+      is_asserted_by do
+        input.attribute('value')&.value&.to_i == @index_form.parameter[param_name]
+      end
+    end
+  end
+
+  it 'max_features入力フォームがあること' do
+    form_id = 'input-index-max_features'
+    label = @html.xpath("#{index_input_xpath}/label[@for='#{form_id}']")
+    is_asserted_by { label.present? }
+    is_asserted_by { label.text.strip == 'max_features' }
+
+    select = @html.xpath("#{index_input_xpath}/select[@id='#{form_id}']")
+    is_asserted_by { select.present? }
+
+    %w[sqrt log2 all].each do |value|
+      is_asserted_by { select.search("option[@value='#{value}']").present? }
+    end
+
+    selected = select.search('option[@selected="selected"]')
+    is_asserted_by do
+      selected.attribute('value')&.value == @index_form.parameter[:max_features]
+    end
+  end
+
+  it '検索ボタンがあること' do
+    button = @html.xpath("#{index_form_xpath}/input[@type='submit']")
+    is_asserted_by { button.present? }
+    is_asserted_by { button.attribute('value').value == '検索' }
   end
 end